+++ examples/account-public-access/versions.tf
@@ -1,10 +1,10 @@
 terraform {
-  required_version = ">= 1.0"
+  required_version = ">= 1.10"
 
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "~> 3.69.0"
+      version = ">= 6.28"
     }
     random = {
       source  = "hashicorp/random"