+++ modules/notification/variables.tf
@@ -16,6 +16,18 @@ variable "create_sqs_policy" {
   default     = true
 }
 
+variable "create_lambda_permission" {
+  description = "Whether to create Lambda permissions or not?"
+  type        = bool
+  default     = true
+}
+
+variable "region" {
+  description = "Region where the resource(s) will be managed. Defaults to the region set in the provider configuration"
+  type        = string
+  default     = null
+}
+
 variable "bucket" {
   description = "Name of S3 bucket to use"
   type        = string